PUT Person fix

+++ src/main/java/home/holymiko/InvestmentScraperApp/Server/Type/DTO/simple/PersonDTO.java
@@ -17,6 +17,6 @@ public class PersonDTO {
     private final String lastName;
 
     private final String email;
-    private final int phone;
+    private final Long phone;
 
 }
